Minor changes to clear_data_output.py and clean_viz_output.py

+++ src/utility_scripts/clear_viz_output.py
@@ -21,11 +21,12 @@ os.chdir(dname)
 folder_path = Path("../../media/figures/raw")
 
 for filename in os.listdir(folder_path):
-    file_path = os.path.join(folder_path, filename)
-    try:
-        if os.path.isfile(file_path) or os.path.islink(file_path):
-            os.unlink(file_path)
-        elif os.path.isdir(file_path):
-            shutil.rmtree(file_path)
-    except Exception as e:
-        print('Failed to delete %s. Reason: %s' % (file_path, e))
+    if filename != ".gitkeep":
+        file_path = os.path.join(folder_path, filename)
+        try:
+            if os.path.isfile(file_path) or os.path.islink(file_path):
+                os.unlink(file_path)
+            elif os.path.isdir(file_path):
+                shutil.rmtree(file_path)
+        except Exception as e:
+            print('Failed to delete %s. Reason: %s' % (file_path, e))
